+++ tests/integ/test_udf.py
@@ -4,6 +4,7 @@
 #
 
 import datetime
+import decimal  # noqa: F401
 import json
 import logging
 import math
@@ -50,7 +51,7 @@ from snowflake.snowpark.exceptions import (
     SnowparkInvalidObjectNameException,
     SnowparkSQLException,
 )
-from snowflake.snowpark.functions import call_udf, col, pandas_udf, parse_json, udf
+from snowflake.snowpark.functions import call_udf, col, lit, pandas_udf, parse_json, udf
 from snowflake.snowpark.types import (
     LTZ,
     NTZ,
@@ -981,6 +982,154 @@ def return_dict(v: dict) -> Dict[str, str]:
 
 @pytest.mark.xfail(
     "config.getoption('local_testing_mode', default=False)",
+    reason="SNOW-1412530 to fix bug",
+    run=False,
+)
+@pytest.mark.parametrize("register_from_file", [True, False])
+def test_register_udf_with_optional_args(session: Session, tmpdir, register_from_file):
+    import_body = """
+import datetime
+import decimal
+from snowflake.snowpark.types import Variant, Geometry, Geography
+from typing import Dict, List, Optional
+"""
+    func_body = """
+def add(x: int = 0, y: int = 0) -> int:
+    return x + y
+
+def snow(x: int = 1) -> Optional[str]:
+    return "snow" if x % 2 == 0 else None
+
+def double_str_list(x: str = "a") -> List[str]:
+    return [x, x]
+
+def return_date(
+    dt: datetime.date = datetime.date(2017, 1, 1)
+) -> datetime.date:
+    return dt
+
+def return_arr(
+    base_arr: List[int], extra_arr: List[int] = [4]
+) -> List[int]:
+    base_arr.extend(extra_arr)
+    return base_arr
+
+def return_all_datatypes(
+    a: int = 1,
+    b: float = 1.0,
+    c: str = "one",
+    d: List[int] = [],
+    e: Dict[str, int] = {"s": 1},
+    f: Variant = {"key": "val"},
+    g: Geometry = "POINT(-122.35 37.55)",
+    h: Geography = "POINT(-122.35 37.55)",
+    i: datetime.datetime = datetime.datetime(2021, 1, 1, 0, 0, 0),
+    j: datetime.date = datetime.date(2021, 1, 1),
+    k: datetime.time = datetime.time(0, 0, 0),
+    l: bytes = b"123",
+    m: bool = True,
+    n: decimal.Decimal = decimal.Decimal(1.0),
+) -> str:
+    final_str = f"{a}, {b}, {c}, {d}, {e}, {f}, {g}, {h}, {i}, {j}, {k}, {l}, {m}, {n}"
+    return final_str
+"""
+    session.add_packages("snowflake-snowpark-python")
+    if register_from_file:
+        file_path = os.path.join(tmpdir, "register_from_file_optional_args.py")
+        with open(file_path, "w") as f:
+            source = f"{import_body}\n{func_body}"
+            f.write(source)
+
+        add_udf = session.udf.register_from_file(file_path, "add")
+        snow_udf = session.udf.register_from_file(file_path, "snow")
+        double_str_list_udf = session.udf.register_from_file(
+            file_path, "double_str_list"
+        )
+        return_date_udf = session.udf.register_from_file(file_path, "return_date")
+        return_arr_udf = session.udf.register_from_file(file_path, "return_arr")
+        return_all_datatypes_udf = session.udf.register_from_file(
+            file_path, "return_all_datatypes"
+        )
+    else:
+        d = {}
+        exec(func_body, {**globals(), **locals()}, d)
+
+        add_udf = session.udf.register(d["add"])
+        snow_udf = session.udf.register(d["snow"])
+        double_str_list_udf = session.udf.register(d["double_str_list"])
+        return_date_udf = session.udf.register(d["return_date"])
+        return_arr_udf = session.udf.register(d["return_arr"])
+        return_all_datatypes_udf = session.udf.register(d["return_all_datatypes"])
+
+    df = session.create_dataframe([[1, 4], [2, 3]]).to_df("a", "b")
+    Utils.check_answer(
+        df.select(
+            add_udf("a", "b"),
+            add_udf("a"),
+            add_udf(),
+        ),
+        [
+            Row(5, 1, 0),
+            Row(5, 2, 0),
+        ],
+    )
+    Utils.check_answer(
+        df.select(
+            snow_udf("a"),
+            snow_udf(),
+        ),
+        [
+            Row(None, None),
+            Row("snow", None),
+        ],
+    )
+    Utils.check_answer(
+        df.select(
+            double_str_list_udf("a"),
+            double_str_list_udf(),
+        ),
+        [
+            Row('[\n  "1",\n  "1"\n]', '[\n  "a",\n  "a"\n]'),
+            Row('[\n  "2",\n  "2"\n]', '[\n  "a",\n  "a"\n]'),
+        ],
+    )
+    Utils.check_answer(
+        df.select(
+            return_date_udf(lit(datetime.date(2024, 4, 4))),
+            return_date_udf(),
+        ),
+        [
+            Row(datetime.date(2024, 4, 4), datetime.date(2017, 1, 1)),
+            Row(datetime.date(2024, 4, 4), datetime.date(2017, 1, 1)),
+        ],
+    )
+    Utils.check_answer(
+        df.select(
+            return_arr_udf(lit([1, 2, 3]), lit(list([4, 5]))),
+            return_arr_udf(lit([1, 2, 3])),
+        ),
+        [
+            Row("[\n  1,\n  2,\n  3,\n  4,\n  5\n]", "[\n  1,\n  2,\n  3,\n  4\n]"),
+            Row("[\n  1,\n  2,\n  3,\n  4,\n  5\n]", "[\n  1,\n  2,\n  3,\n  4\n]"),
+        ],
+    )
+
+    default_row = "1, 1.0, one, [], {'s': 1}, {'key': 'val'}, {'coordinates': [-122.35, 37.55], 'type': 'Point'}, {'coordinates': [-122.35, 37.55], 'type': 'Point'}, 2021-01-01 00:00:00, 2021-01-01, 00:00:00, b'123', True, 1.000000000000000000"
+    custom_row = "2, 2.0, two, [], {'s': 1}, {'key': 'val'}, {'coordinates': [-122.35, 37.55], 'type': 'Point'}, {'coordinates': [-122.35, 37.55], 'type': 'Point'}, 2021-01-01 00:00:00, 2021-01-01, 00:00:00, b'123', True, 1.000000000000000000"
+    Utils.check_answer(
+        df.select(
+            return_all_datatypes_udf(),
+            return_all_datatypes_udf(lit(2), lit(2.0), lit("two")),
+        ),
+        [
+            Row(default_row, custom_row),
+            Row(default_row, custom_row),
+        ],
+    )
+
+
+@pytest.mark.xfail(
+    "config.getoption('local_testing_mode', default=False)",
     reason="Database objects are session scoped in Local Testing",
     run=False,
 )
@@ -1074,7 +1223,7 @@ def test_udf_negative(session, local_testing_mode):
     with pytest.raises(ValueError) as ex_info:
         udf1("a", "")
     assert (
-        "Incorrect number of arguments passed to the UDF: Expected: 1, Found: 2"
+        "Incorrect number of arguments passed to the UDF: Expected: <=1, Found: 2"
         in str(ex_info)
     )